requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
ollama==0.1.7
google-generativeai==0.3.2
python-dotenv==1.0.0

